Import of GTERMARS/Alien-Prototype-Window-1.3.2 from CPAN.
gitpan-cpan-distribution: Alien-Prototype-Window
gitpan-cpan-version:      1.3.2
gitpan-cpan-path:         GTERMARS/Alien-Prototype-Window-1.3.2.tar.gz
gitpan-cpan-author:       GTERMARS
gitpan-cpan-maturity:     released

+++ t/02-install.t
@@ -1,21 +1,38 @@
 use strict;
 use warnings;
-use Test::More tests => 5;
+use Test::More;
 use Test::Exception;
 use File::Path qw(rmtree);
 use Alien::Prototype::Window;
 
 my $dir = 't/eraseme';
 
-# Do an install and make sure that at least one file from each of the 'lib' and
-# 'src' directories was installed properly.
-Alien::Prototype::Window->install( $dir );
-foreach my $file (qw( window.js window_ext.js window_effects.js tooltip.js )) {
-    ok( -e "$dir/$file", "$dir/$file exists" );
+# Figure out how many tests we're going to run
+my @files = Alien::Prototype::Window->files();
+plan tests => 2 + scalar @files;
+
+# Clean up from any previous test run.
+cleanup_old_test_run: {
+    rmtree( $dir ) if (-e $dir);
 }
 
-# Re-install into the same directory, to make sure that it doesn't choke.
-lives_ok { Alien::Prototype::Window->install($dir) };
+# Install, and make sure that all of the files got installed properly.
+install_pwc: {
+    Alien::Prototype::Window->install( $dir );
+    # make sure script.aculo.us got installed
+    ok( -e "$dir/scriptaculous.js", "script.aculo.us installed" );
+    # make sure our files got installed
+    foreach my $file (@files) {
+        ok( -e "$dir/$file", "$file exists" );
+    }
+}
 
-# Clean out the test directory
-rmtree( $dir );
+# Install on top of an existing install; shouldn't choke
+reinstall_pwc: {
+    lives_ok { Alien::Prototype::Window->install($dir) };
+}
+
+# Clean up after ourselves
+cleanup: {
+    rmtree( $dir );
+}
